version 0.0.0.3, make more intuitive
* Using coroutines for describing the game
* Introduced Picture like gloss
* Moved Data.Bullet, Data.Danmaku -> Control.Monad.Bullet,
Control.Monad.Danmaku
* Added tinny SEs
* And reduced a lot of codes

+++ MonadicShooter/Field.hs
@@ -1,16 +1,16 @@
 module MonadicShooter.Field where
-import Data.Vect
+import Data.Vect.Double
 
-upperBound :: Float
+upperBound :: Double
 upperBound = 0.0
 
-leftBound :: Float
+leftBound :: Double
 leftBound = 0.0
 
-lowerBound :: Float
+lowerBound :: Double
 lowerBound = 480.0
 
-rightBound :: Float
+rightBound :: Double
 rightBound = 480.0
 
 upperLeft :: Vec2
